fix: alterando tipagens e construtores de selects

+++ app-backend-atelie/src/main/java/com/eseltech/appbackendatelie/DTO/dash/MaterialEstoqueDTO.java
@@ -10,5 +10,5 @@ public record MaterialEstoqueDTO(
         BigDecimal preco,
         String categoria,
         BigDecimal consumoUltimoMes,
-        Double margemEstoquePercentual
+        BigDecimal margemEstoquePercentual
 ) {}
